refactory: 코틀린 파일 정리
