+++ app/src/main/java/umc/mobile/project/login/MyCustomDialogInterface.kt
@@ -1,4 +1,4 @@
-package umc.mobile.project
+package umc.mobile.project.login
 
 interface MyCustomDialogInterface {
     fun onbtnGotoMainClicked()
