+++ app/src/main/java/umc/mobile/project/news/NewsUpLoadFragment.kt
@@ -1,4 +1,4 @@
-package umc.mobile.project
+package umc.mobile.project.news
 
 
 import android.os.Bundle
@@ -7,6 +7,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
+import umc.mobile.project.R
 import umc.mobile.project.databinding.FragmentNewsUploadBinding
