+++ app/src/main/java/umc/mobile/project/news/UpLoadDataRVAdapter.kt
@@ -1,4 +1,4 @@
-package umc.mobile.project
+package umc.mobile.project.news
 
 import android.util.SparseBooleanArray
 import android.view.LayoutInflater
@@ -24,7 +24,7 @@ class UpLoadDataRVAdapter(private val UpLoadDataList: ArrayList<UpLoadData>) : R
     }
 
     //ViewHolder 만들어질 때 실행할 동작
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): UpLoadDataRVAdapter.DataViewHolder {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): DataViewHolder {
         val viewBinding =
             ItemNewsUploadDataBinding.inflate(LayoutInflater.from(parent.context))
         return DataViewHolder(viewBinding)
