+++ app/src/main/java/umc/mobile/project/signup/SignUpAlarmKeywordActivity.kt
@@ -1,27 +1,29 @@
-package umc.mobile.project
+package umc.mobile.project.signup
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import umc.mobile.project.databinding.ActivitySignUpRegionBinding
+import umc.mobile.project.databinding.ActivitySignUpAlarmKeywordBinding
 
-class SignUpRegionActivity : AppCompatActivity() {
-    private lateinit var viewBinding: ActivitySignUpRegionBinding
+class SignUpAlarmKeywordActivity : AppCompatActivity() {
+    private lateinit var viewBinding: ActivitySignUpAlarmKeywordBinding
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        viewBinding = ActivitySignUpRegionBinding.inflate(layoutInflater)
+        viewBinding = ActivitySignUpAlarmKeywordBinding.inflate(layoutInflater)
         setContentView(viewBinding.root)
 
         viewBinding.btnNext.setOnClickListener {
-            val intent = Intent(this, SignUpUserCategoryActivity::class.java)
+            val intent = Intent(this, SignUpRegionActivity::class.java)
             startActivity(intent)
         }
 
         viewBinding.btnBack.setOnClickListener {
-            val intent = Intent(this, SignUpAlarmKeywordActivity::class.java)
+            val intent = Intent(this, SignUpActivity::class.java)
             finish()
             startActivity(intent)
 
         }
     }
+
+
 }
