+++ app/src/main/java/umc/mobile/project/signup/SignUpRegionActivity.kt
@@ -1,29 +1,27 @@
-package umc.mobile.project
+package umc.mobile.project.signup
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import umc.mobile.project.databinding.ActivitySignUpAlarmKeywordBinding
+import umc.mobile.project.databinding.ActivitySignUpRegionBinding
 
-class SignUpAlarmKeywordActivity : AppCompatActivity() {
-    private lateinit var viewBinding: ActivitySignUpAlarmKeywordBinding
+class SignUpRegionActivity : AppCompatActivity() {
+    private lateinit var viewBinding: ActivitySignUpRegionBinding
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        viewBinding = ActivitySignUpAlarmKeywordBinding.inflate(layoutInflater)
+        viewBinding = ActivitySignUpRegionBinding.inflate(layoutInflater)
         setContentView(viewBinding.root)
 
         viewBinding.btnNext.setOnClickListener {
-            val intent = Intent(this, SignUpRegionActivity::class.java)
+            val intent = Intent(this, SignUpUserCategoryActivity::class.java)
             startActivity(intent)
         }
 
         viewBinding.btnBack.setOnClickListener {
-            val intent = Intent(this, SignUpActivity::class.java)
+            val intent = Intent(this, SignUpAlarmKeywordActivity::class.java)
             finish()
             startActivity(intent)
 
         }
     }
-
-
 }
