+++ app/src/main/res/layout/activity_news.xml
@@ -5,7 +5,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".NewsActivity">
+    tools:context=".news.NewsActivity">
 
     <include
         android:id="@+id/topView"
