+++ app/src/main/res/layout/activity_sign_up_region.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SignUpAlarmKeywordActivity"
+    tools:context=".signup.SignUpAlarmKeywordActivity"
     android:orientation="vertical"
     >
