+++ app/src/main/res/layout/activity_sign_up_user_personal_info.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".SignUpAlarmKeywordActivity">
+    tools:context=".signup.SignUpAlarmKeywordActivity">
 
     <Button
         android:id="@+id/btnBack"
